+++ packages/cli/CHANGELOG.md
@@ -41,4 +41,4 @@
 
 ### Minor Changes
 
-- [`4b671bb`](https://github.com/good-jinu/chatcore/commit/4b671bb4c12719dd27595053e1af95e4e70abc9b) Thanks [@good-jinu](https://github.com/good-jinu)! - Add a separate ChatCore CLI package with `chatcore schema generate` for SQL schema generation.
+- [`4b671bb`](https://github.com/good-jinu/chatcore/commit/4b671bb4c12719dd27595053e1af95e4e70abc9b) Thanks [@good-jinu](https://github.com/good-jinu)! - Add a separate MessageWeave CLI package with `messageweave schema generate` for SQL schema generation.